==> 02_etl_dataframe_relatorio_vendas.ipynb <==
# Relatório de Vendas DataFiniti Eletronics
**descrição das colunas do dataframe e breve explicação sobre o mesmo**<br>


Neste notebook iremos explorar o dataframe de vendas e concatená-lo com o calendário


#### Imports

import pandas as pd
import numpy as np
