==> 07_model_holiday_summer.ipynb <==
#### Imports


import statsmodels.api as sm
import pandas as pd
import numpy as np


#### Carregando os arquivos


df_relatorio_vendas = (
    pd.read_csv('./dataset/relatorio_vendas_etapa_6.csv', sep=';')
)


#### Preparação dos dados


**Iremos inicialmente separar o relatório em duas parte:**

1. 
2. Vendas em dias comuns.<br>
   **De posse destes resultaods, iremos filtra cada dataframe utilizando a coluna IN_IS_SUMMER**<br>
3. Vendas em dias de feriado + verão<br>
4. Vendas em dias comuns + verão<br>

**Próximo passo será separar para cada datframe os 25 campeões de venda**


# Separando as Vendas em dias de feriado.
df_holiday = (
    df_relatorio_vendas[
        df_relatorio_vendas['IN_IS_HOLIDAY'] == 1
    ]
)

# Separando as Vendas em dias de feriado + verão.
df_summer_holiday = (
    df_holiday[
        df_holiday['IN_IS_SUMMER'] == 1
    ]
)

# Separando as Vendas em dias comuns.
df_non_holiday = (
    df_relatorio_vendas[
        df_relatorio_vendas['IN_IS_HOLIDAY'] == 0
    ]
)

# Separando as Vendas em dias de comuns + verão.
df_summer_non_holiday = (
    df_non_holiday[
        df_non_holiday['IN_IS_SUMMER'] == 1
    ]
)


#### Separando os 25 campeões de vendas.


# Mostrando a quantidade vendida por todos os produtos
(
    df_relatorio_vendas
    .groupby(
        [
            'CD_PRODUCT_COLOR'
        ]
    )
    .agg(
        DATA_VOLUME=(
            'DT_ORDER_DATE', 'count'
        )
    )
    .sort_values(
        'DATA_VOLUME'
        , ascending=False
    )
)



# Separando os 25 campeões de vendas
top_25_products = (
    df_relatorio_vendas
    .groupby('CD_PRODUCT_COLOR')
    .agg(
        VL_TOTAL_SOLD=(
            'QT_SOLD', 'sum'
        )
    )
    .sort_values('VL_TOTAL_SOLD', ascending=False)
    .head(25)
    .index
)


# Mostrando a quantidade de vendas dos 25 campeões de venda.
(
    df_relatorio_vendas[
        df_relatorio_vendas['CD_PRODUCT_COLOR']
        .isin(top_25_products)
    ]
    .groupby(
        [
            'CD_PRODUCT_COLOR'
        ]
    )
    .agg(
        DATA_VOLUME=(
            'DT_ORDER_DATE', 'count'
        )
    )
    .sort_values('DATA_VOLUME', ascending=False)
)


#### Filtrando os dataframes por cenário

df_holiday = (
    df_holiday[
        df_holiday['CD_PRODUCT_COLOR']
        .isin(top_25_products)
    ]
)

df_summer_holiday = (
    df_summer_holiday[
        df_summer_holiday['CD_PRODUCT_COLOR']
        .isin(top_25_products)
    ]
)

df_non_holiday = (
    df_non_holiday[
        df_non_holiday['CD_PRODUCT_COLOR']
        .isin(top_25_products)
    ]
)

df_summer_non_holiday = (
    df_summer_non_holiday[
        df_summer_non_holiday['CD_PRODUCT_COLOR']
        .isin(top_25_products)
    ]
)


# Searching for the sold amount for each product
df_sold_amount = (df_relatorio_vendas
                  .groupby('CD_PRODUCT_COLOR', as_index=False)
                  .agg(QT_TOTAL_SOLD_AMOUNT=('QT_SOLD', 'sum'))
                  ).sort_values('QT_TOTAL_SOLD_AMOUNT', ascending=False)


#### Modelos


**`Vendas em feriados`**    

results_values = {
    "CD_PRODUCT_COLOR": [],
    "VL_PRICE_ELASTICITY": [],
    "NR_P_VALUE": [],
}

# Append x_values with y_values per same product name
for product in df_holiday['CD_PRODUCT_COLOR'].unique():
    df = df_holiday[(df_holiday['CD_PRODUCT_COLOR'] == product)].reset_index()

    # Linear Regression Model
    x_value = df['VL_DISCOUNT_PRICE_A2'].values
    y_value = df['QT_SOLD'].values
    X = sm.add_constant(x_value, has_constant='add')

    model = sm.OLS(y_value, X)
    result = model.fit()
    ypred = result.predict(X)

    rsquared = result.rsquared
    coefficient_pvalue = result.f_pvalue
    intercept, slope = result.params
    mean_price = np.mean(x_value)
    mean_quantity = np.mean(y_value)

    if coefficient_pvalue <= 0.05:

        # Price elasticity Formula
        price_elasticity = (slope)*(mean_price/mean_quantity)

        # Append results into dictionary for dataframe
        results_values["CD_PRODUCT_COLOR"].append(product)
        results_values["VL_PRICE_ELASTICITY"].append(price_elasticity)
        results_values["NR_P_VALUE"].append(coefficient_pvalue)


df_OLS_price_elasticity = pd.DataFrame.from_dict(results_values)

df_OLS_price_elasticity = (
    df_OLS_price_elasticity[[
        'CD_PRODUCT_COLOR', 'VL_PRICE_ELASTICITY', 'NR_P_VALUE'
    ]]
)

df_OLS_price_elasticity

**`Vendas em feriados + verão`**    

results_values = {
    "CD_PRODUCT_COLOR": [],
    "VL_PRICE_ELASTICITY": [],
    "NR_P_VALUE": [],
}

# Append x_values with y_values per same product name
for product in df_summer_holiday['CD_PRODUCT_COLOR'].unique():
    df = df_summer_holiday[(df_summer_holiday['CD_PRODUCT_COLOR'] == product)].reset_index()

    # Linear Regression Model
    x_value = df['VL_DISCOUNT_PRICE_A2'].values
    y_value = df['QT_SOLD'].values
    X = sm.add_constant(x_value, has_constant='add')

    model = sm.OLS(y_value, X)
    result = model.fit()
    ypred = result.predict(X)

    rsquared = result.rsquared
    coefficient_pvalue = result.f_pvalue
    intercept, slope = result.params
    mean_price = np.mean(x_value)
    mean_quantity = np.mean(y_value)

    if coefficient_pvalue <= 0.05:

        # Price elasticity Formula
        price_elasticity = (slope)*(mean_price/mean_quantity)

        # Append results into dictionary for dataframe
        results_values["CD_PRODUCT_COLOR"].append(product)
        results_values["VL_PRICE_ELASTICITY"].append(price_elasticity)
        results_values["NR_P_VALUE"].append(coefficient_pvalue)


df_OLS_price_elasticity = pd.DataFrame.from_dict(results_values)

df_OLS_price_elasticity = (
    df_OLS_price_elasticity[[
        'CD_PRODUCT_COLOR', 'VL_PRICE_ELASTICITY', 'NR_P_VALUE'
    ]]
)

df_OLS_price_elasticity

**`Vendas em dias comuns`**    

results_values = {
    "CD_PRODUCT_COLOR": [],
    "VL_PRICE_ELASTICITY": [],
    "NR_P_VALUE": [],
}

# Append x_values with y_values per same product name
for product in df_non_holiday['CD_PRODUCT_COLOR'].unique():
    df = df_non_holiday[(df_non_holiday['CD_PRODUCT_COLOR']
                         == product)].reset_index()

    # Linear Regression Model
    x_value = df['VL_DISCOUNT_PRICE_A2'].values
    y_value = df['QT_SOLD'].values
    X = sm.add_constant(x_value, has_constant='add')

    model = sm.OLS(y_value, X)
    result = model.fit()
    ypred = result.predict(X)

    rsquared = result.rsquared
    coefficient_pvalue = result.f_pvalue
    intercept, slope = result.params
    mean_price = np.mean(x_value)
    mean_quantity = np.mean(y_value)

    if coefficient_pvalue <= 0.05:

        # Price elasticity Formula
        price_elasticity = (slope)*(mean_price/mean_quantity)

        # Append results into dictionary for dataframe
        results_values["CD_PRODUCT_COLOR"].append(product)
        results_values["VL_PRICE_ELASTICITY"].append(price_elasticity)
        results_values["NR_P_VALUE"].append(coefficient_pvalue)


df_OLS_price_elasticity = pd.DataFrame.from_dict(results_values)

df_OLS_price_elasticity = (
    df_OLS_price_elasticity[[
        'CD_PRODUCT_COLOR', 'VL_PRICE_ELASTICITY', 'NR_P_VALUE'
    ]]
)

df_OLS_price_elasticity

**`Vendas em dias comuns + verão`**    

results_values = {
    "CD_PRODUCT_COLOR": [],
    "VL_PRICE_ELASTICITY": [],
    "NR_P_VALUE": [],
}

# Append x_values with y_values per same product name
for product in df_summer_non_holiday['CD_PRODUCT_COLOR'].unique():
    df = df_summer_non_holiday[(
        df_summer_non_holiday['CD_PRODUCT_COLOR'] == product)].reset_index()

    # Linear Regression Model
    x_value = df['VL_DISCOUNT_PRICE_A2'].values
    y_value = df['QT_SOLD'].values
    X = sm.add_constant(x_value, has_constant='add')

    model = sm.OLS(y_value, X)
    result = model.fit()
    ypred = result.predict(X)

    rsquared = result.rsquared
    coefficient_pvalue = result.f_pvalue
    intercept, slope = result.params
    mean_price = np.mean(x_value)
    mean_quantity = np.mean(y_value)

    if coefficient_pvalue <= 0.05:

        # Price elasticity Formula
        price_elasticity = (slope)*(mean_price/mean_quantity)

        # Append results into dictionary for dataframe
        results_values["CD_PRODUCT_COLOR"].append(product)
        results_values["VL_PRICE_ELASTICITY"].append(price_elasticity)
        results_values["NR_P_VALUE"].append(coefficient_pvalue)


df_OLS_price_elasticity = pd.DataFrame.from_dict(results_values)

df_OLS_price_elasticity = (
    df_OLS_price_elasticity[[
        'CD_PRODUCT_COLOR', 'VL_PRICE_ELASTICITY', 'NR_P_VALUE'
    ]]
)

df_OLS_price_elasticity
